allowed to save and load Axis and Group objects in HDF files (#578) : - added to_hdf method to Axis and Group - updated read_hdf (inout/hdf.py) - updated documentation of Session's methods - updated doctests of Session.load and Session.save - added context manager LHDFStore (utils/misc.py)
refactored package inout: created one module per file extension or external object type like in pandas/io/:

new modules:
  - common.py
  - pandas.py
  - csv.py
  - excel.py
  - hdf.py
  - sas.py
  - misc.py
  - pickle.py

renamed modules:
  - excel.py --> xw_excel.py

deleted modules:
  - array.py
