+++ larray/inout/__init__.py
@@ -1,5 +1,9 @@
 from __future__ import absolute_import, division, print_function
 
+from larray.inout.pandas import *
+from larray.inout.csv import *
+from larray.inout.misc import *
 from larray.inout.excel import *
-from larray.inout.array import *
-from larray.inout.session import *
+from larray.inout.hdf import *
+from larray.inout.sas import *
+from larray.inout.xw_excel import *
